LPD-52709 Move everything to the same package
