+++ modules/apps/frontend-js/frontend-js-web/src/main/java/com/liferay/frontend/js/web/internal/hashed/files/request/helper/StaticFileRequestHelperInfo.java
@@ -5,8 +5,6 @@
 
 package com.liferay.frontend.js.web.internal.hashed.files.request.helper;
 
-import com.liferay.frontend.js.web.internal.hashed.files.request.RequestHelperInfo;
-
 /**
  * @author Iván Zaera Avellón
  */
